git-svn-id: https://subversion.imagemagick.org/subversion/ImageMagick/branches/ImageMagick-6.7.5@6802 aa41f4f7-0bf4-0310-aa73-e5a19afd5a74

+++ wand/mogrify.c
@@ -182,7 +182,7 @@ WandExport MagickBooleanType MagickCommandGenesis(ImageInfo *image_info,
     TimerInfo
       *timer;
 
-    SetOpenMPMaximumThreads(n);
+    SetOpenMPMaximumThreads((int) n);
     timer=AcquireTimerInfo();
     if (concurrent == MagickFalse)
       {
